+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - List all MDM servers in an organisation
 - Show all device IDs assigned to a specific MDM server
 - Show which MDM server a device is assigned to
+- Show the ID of the MDM server a device is assigned to
+- Show information about the MDM server a device is assigned to
 
 ## 0.1.2 - 2025-07-14